+++ src/config/battleConfig.js
@@ -1069,6 +1069,54 @@ const effectConfig = {
             battle.eventHandler.unregisterListener(listenerId);
         }
     },
+    "moneyBags": {
+        "name": "Money Bags",
+        "description": "When the user takes damage, heal other allies by half the damage taken.",
+        "type": effectTypes.BUFF,
+        "dispellable": true,
+        "effectAdd": function(battle, source, target) {
+            battle.addToLog(`${source.name} is sharing the wealth!`);
+            const listener = {
+                initialArgs: {
+                    pokemon: source,
+                },
+                execute: function(initialArgs, args) {
+                    const moneyBagsPokemon = initialArgs.pokemon;
+                    if (moneyBagsPokemon.isFainted) {
+                        return;
+                    }
+                    const damagedPokemon = args.target;
+                    if (moneyBagsPokemon !== damagedPokemon) {
+                        return;
+                    }
+
+                    const damage = args.damage;
+                    const allyParty = moneyBagsPokemon.battle.parties[moneyBagsPokemon.teamName];
+                    const moneyBagsAllies = moneyBagsPokemon.getPatternTargets(
+                        allyParty,
+                        targetPatterns.ALL_EXCEPT_SELF,
+                        moneyBagsPokemon.position,
+                    );
+
+                    moneyBagsPokemon.battle.addToLog(`Coins drop from ${moneyBagsPokemon.name}'s wallet!`);
+                    for (const ally of moneyBagsAllies) {
+                        const healAmount = Math.floor(damage / 2);
+                        moneyBagsPokemon.giveHeal(healAmount, ally, {
+                            "type": "moneyBags",
+                        });
+                    }
+                }
+            }
+            const listenerId = battle.eventHandler.registerListener(battleEventNames.AFTER_DAMAGE_TAKEN, listener);
+            return {
+                "listenerId": listenerId,
+            };
+        },
+        "effectRemove": function(battle, target, args) {
+            const listenerId = args.listenerId;
+            battle.eventHandler.unregisterListener(listenerId);
+        }
+    },
     "statusImmunity": {
         "name": "Status Immunity",
         "description": "The target is immune to status conditions.",
@@ -2677,6 +2725,19 @@ const moveConfig = {
         "damageType": damageTypes.OTHER,
         "description": "The user emits a screech harsh enough to sharply lower the targets' Defense for 2 turns.",
     },
+    "m105": {
+        "name": "Recover",
+        "type": types.NORMAL,
+        "power": null,
+        "accuracy": null,
+        "cooldown": 3,
+        "targetType": targetTypes.ALLY,
+        "targetPosition": targetPositions.SELF,
+        "targetPattern": targetPatterns.SINGLE,
+        "tier": moveTiers.POWER,
+        "damageType": damageTypes.OTHER,
+        "description": "The user recovers 50% of its max HP.",
+    },
     "m106": {
         "name": "Harden",
         "type": types.NORMAL,
@@ -3030,6 +3091,19 @@ const moveConfig = {
         "tier": moveTiers.ULTIMATE,
         "damageType": damageTypes.PHYSICAL,
         "description": "A consecutive spinning three-kick attack. Deals damage 3 times.",
+    },
+    "m168": {
+        "name": "Thief",
+        "type": types.DARK,
+        "power": 60,
+        "accuracy": 100,
+        "cooldown": 4,
+        "targetType": targetTypes.ENEMY,
+        "targetPosition": targetPositions.ANY,
+        "targetPattern": targetPatterns.SINGLE,
+        "tier": moveTiers.POWER,
+        "damageType": damageTypes.PHYSICAL,
+        "description": "The user attacks and steals the target's buffs simultaneously.",
     },
     "m175": {
         "name": "Flail",
@@ -3510,7 +3584,7 @@ const moveConfig = {
         "targetPattern": targetPatterns.SINGLE,
         "tier": moveTiers.POWER,
         "damageType": damageTypes.OTHER,
-        "description": "The user creates a reflective field for 1 turn. When dmaged by a Special move, deals damage back to the target with 1.5x base power.",
+        "description": "The user creates a reflective field for 1 turn. When damaged by a Special move, deals damage back to the target with 1.5x base power.",
     },
     "m245": {
         "name": "Extreme Speed",
@@ -4971,6 +5045,32 @@ const moveConfig = {
         "damageType": damageTypes.SPECIAL,
         "description": "The user fires a beam of dark energy at the target, dealing damage and dispelling all buffs.",
     },
+    "m20007": {
+        "name": "Scammed!",
+        "type": types.DARK,
+        "power": null,
+        "accuracy": null,
+        "cooldown": 6,
+        "targetType": targetTypes.ENEMY,
+        "targetPosition": targetPositions.ANY,
+        "targetPattern": targetPatterns.ALL,
+        "tier": moveTiers.ULTIMATE,
+        "damageType": damageTypes.OTHER,
+        "description": "The user makes an unfair deal with the enemy team, increasing enemies combat readiness by 10%, in exchange for increasing all ally combat readiness by 30%.",
+    },
+    "m20008": {
+        "name": "All-In",
+        "type": types.GHOST,
+        "power": null,
+        "accuracy": null,
+        "cooldown": 5,
+        "targetType": targetTypes.ALLY,
+        "targetPosition": targetPositions.SELF,
+        "targetPattern": targetPatterns.ALL,
+        "tier": moveTiers.ULTIMATE,
+        "damageType": damageTypes.OTHER,
+        "description": "The user goes all-in, sacrificing all its wealth, sharply reducing its defenses for 2 turns and reducing its health by 50%. In exchange, increases the duration of all ally buffs by 2 turns.",
+    },
 };
 
 const moveExecutes = {
@@ -5864,6 +5964,18 @@ const moveExecutes = {
             }
         }
     },
+    "m105": function (battle, source, primaryTarget, allTargets, missedTargets) {
+        const moveId = "m105";
+        const moveData = moveConfig[moveId];
+        for (const target of allTargets) {
+            // heal 50%
+            const healAmount = Math.floor(target.maxHp * 0.5);
+            source.giveHeal(healAmount, target, {
+                type: "move",
+                moveId: moveId
+            });
+        }
+    },
     "m106": function (battle, source, primaryTarget, allTargets, missedTargets) {
         const moveId = "m106";
         const moveData = moveConfig[moveId];
@@ -6304,6 +6416,40 @@ const moveExecutes = {
                 type: "move",
                 moveId: moveId
             });
+        }
+    },
+    "m168": function (battle, source, primaryTarget, allTargets, missedTargets) {
+        const moveId = "m168";
+        const moveData = moveConfig[moveId];
+        for (const target of allTargets) {
+            const miss = missedTargets.includes(target);
+            const damageToDeal = calculateDamage(moveData, source, target, miss);
+            source.dealDamage(damageToDeal, target, {
+                type: "move",
+                moveId: moveId
+            });
+
+            // if not miss, attempt to steal all buffs
+            if (!miss) {
+                const possibleBuffs = Object.keys(target.effectIds).filter(effectId => {
+                    const effectData = effectConfig[effectId];
+                    return effectData.type === effectTypes.BUFF && effectData.dispellable;
+                });
+                if (possibleBuffs.length == 0) {
+                    return;
+                }
+
+                for (const buffIdToSteal of possibleBuffs) {
+                    const buffToSteal = target.effectIds[buffIdToSteal];
+                    // steal buff
+                    const dispelled = target.dispellEffect(buffIdToSteal);
+                    if (!dispelled) {
+                        continue;
+                    }
+                    // apply buff to self
+                    source.addEffect(buffIdToSteal, buffToSteal.duration, buffToSteal.source, buffToSteal.initialArgs);
+                }
+            }
         }
     },
     "m175": function (battle, source, primaryTarget, allTargets, missedTargets) {
@@ -8894,6 +9040,41 @@ const moveExecutes = {
             }
         }
     },
+    "m20007": function (battle, source, primaryTarget, allTargets, missedTargets) {
+        const moveId = "m20007";
+        const moveData = moveConfig[moveId];
+        // get all non-fainted, hitable pokemon
+        const targets = Object.values(battle.allPokemon).filter(p => battle.isPokemonHittable(p, moveId));
+        for (const target of targets) {
+            const crBoost = target.teamName === source.teamName ? 30 : 10;
+            target.boostCombatReadiness(source, crBoost);
+        }
+    },
+    "m20008": function (battle, source, primaryTarget, allTargets, missedTargets) {
+        const moveId = "m20008";
+        const moveData = moveConfig[moveId];
+        for (const target of allTargets) {
+            // if primary, reduce hp by 50% and apply greater def down, spd down 2 turns
+            if (target === primaryTarget) {
+                source.dealDamage(Math.floor(target.hp / 2), target, {
+                    type: "move",
+                    moveId: moveId
+                });
+
+                target.addEffect("greaterDefDown", 2, source);
+                target.addEffect("greaterSpdDown", 2, source);
+            }
+
+            // increase all buff durations by 2 if dispellable
+            battle.addToLog(`${target.name} is invigorated!`);
+            for (const effectId of Object.keys(target.effectIds)) {
+                const effectData = effectConfig[effectId];
+                if (effectData.type === effectTypes.BUFF && effectData.dispellable) {
+                    target.effectIds[effectId].duration += 2;
+                }
+            }
+        }
+    },
 };
 
 const abilityConfig = {
@@ -10996,6 +11177,43 @@ const abilityConfig = {
             battle.eventHandler.unregisterListener(abilityData.listenerId);
         }
     },
+    "198": {
+        "name": "Stakeout",
+        "description": "Allies deal 30% more damage to enemies with greater than 70% combat readiness.",
+        "abilityAdd": function (battle, source, target) {
+            const listener = {
+                initialArgs: {
+                    pokemon: target,
+                },
+                execute: function(initialArgs, args) {
+                    const sourcePokemon = args.source;
+                    const targetPokemon = args.target;
+                    if (initialArgs.pokemon.teamName === targetPokemon.teamName) {
+                        return;
+                    }
+
+                    if (targetPokemon.combatReadiness < 75) {
+                        return;
+                    }
+
+                    args.damage = Math.round(args.damage * 1.3);
+                    targetPokemon.battle.addToLog(`${initialArgs.pokemon.name}'s Stakeout increases the damage!`);
+                }
+            }
+            const listenerId = battle.eventHandler.registerListener(battleEventNames.BEFORE_DAMAGE_DEALT, listener);
+            return {
+                "listenerId": listenerId,
+            };
+        },
+        "abilityRemove": function (battle, source, target) {
+            const ability = target.ability;
+            if (!ability || ability.abilityId !== "198" || !ability.data) {
+                return;
+            }
+            const abilityData = ability.data;
+            battle.eventHandler.unregisterListener(abilityData.listenerId);
+        }
+    },
     "20001": {
         "name": "Mind Presence",
         "description": "Whenever an enemy ends a turn, increase the user's combat readiness by 10% attacking stats by 2% without triggering effects. The user cannot be damaged by more than 35% of its max HP at a time. The user is immune to instant-faint effects.",
@@ -11518,6 +11736,35 @@ const abilityConfig = {
             target.spa -= Math.floor(target.batk * 0.25);
         }
     },
+    "20010": {
+        "name": "Money Bags",
+        "description": "At the start of the battle, the user gains the Money Bags buff, which heals allies when taking damage.",
+        "abilityAdd": function (battle, source, target) {
+            const listener = {
+                initialArgs: {
+                    pokemon: target,
+                },
+                execute: function(initialArgs, args) {
+                    const pokemon = initialArgs.pokemon;
+                    // add moneybags buff
+                    pokemon.addEffect("moneyBags", 1, pokemon);
+                }
+            };
+
+            const listenerId = battle.eventHandler.registerListener(battleEventNames.BATTLE_BEGIN, listener);
+            return {
+                listenerId: listenerId,
+            }
+        },
+        "abilityRemove": function (battle, source, target) {
+            const ability = target.ability;
+            if (!ability || ability.abilityId !== "m20010" || !ability.data) {
+                return;
+            }
+            const abilityData = ability.data;
+            battle.eventHandler.unregisterListener(abilityData.listenerId);
+        }
+    }
 };
 
 module.exports = {